+++ CHANGELOG.rst
@@ -2,6 +2,14 @@
 CHANGELOG
 *********
 
+
+1.0.1-alpha
+-----------
+
+
+* csvnorm: renamed and refactored from csvsqueeze
+
+
 1.0.0-alpha
 -----------
 
@@ -11,6 +19,6 @@ Basic functionality and testing for these tools:
 * csvflatten
 * csvsed (missing ``--replace/--whole``)
 * csvslice
-* csvsqueeze (likely to berefactored)
+* csvsqueeze (likely to be refactored)